perf: improve sort() (#37)
* improve sort()

* lint, add comment

* lint again

* fix comment

* lint

* add barebones test setup

* fix implementation

* fix

* update comment

* update comment

* requested changes

* rm unnecessary assignment

* improve tests
