+++ src/utils.js
@@ -25,7 +25,25 @@ function createUndiciRequestOptions (request, signal) {
   }
 }
 
+// Sorts 1-dimensional headers array by traversing it, from the end, and
+// comparing each pair to the previous pair. If the `>` comparison for
+// the previous pair is true, remove previous pair, move it to the end,
+// and reset the counter so it can restart the traversal and not miss anything
+function sort1d (arr) {
+  let i = arr.length
+  while (i > 0) {
+    if (arr[i - 2] > arr[i]) {
+      arr.push(...arr.splice(i - 2, 2))
+      i = arr.length
+    } else {
+      i -= 2
+    }
+  }
+  return arr
+}
+
 module.exports = {
+  sort1d,
   isReadable,
   AbortError,
   createUndiciRequestOptions
