+++ test/utils.js
@@ -5,6 +5,7 @@ const stream = require('stream')
 const Request = require('../src/request')
 
 const {
+  sort1d,
   isReadable,
   createUndiciRequestOptions
 } = require('../src/utils')
@@ -40,5 +41,81 @@ tap.test('createUndiciRequestOptions', t => {
     }
 
     t.strictDeepEqual(found, wanted)
+  })
+})
+
+tap.test('sort1d', t => {
+  t.plan(1)
+
+  t.test('correctly sorts 1-dimensional header arrays', t => {
+    t.plan(2)
+
+    function legacySort1d (headers) {
+      const namesAndOriginalIndex = []
+      // O(n)
+      for (let i = 0; i < headers.length; i += 2) {
+        namesAndOriginalIndex.push([headers[i], i])
+      }
+      // O(n log n)
+      namesAndOriginalIndex.sort((a, b) => {
+        const nameA = a[0]
+        const nameB = b[0]
+        if (nameA < nameB) {
+          return -1
+        } else if (nameA > nameB) {
+          return 1
+        } else {
+          return 0
+        }
+      })
+      const sorted = []
+      // O(n/2)
+      for (const [name, index] of namesAndOriginalIndex) {
+        sorted.push(name)
+        sorted.push(headers[index + 1])
+      }
+      return sorted
+    }
+
+    const wanted = [
+      'header-1',
+      'value',
+      'header-5',
+      'value',
+      'header-a',
+      'value',
+      'header-b',
+      'value',
+      'header-c',
+      'value',
+      'header-d',
+      'value',
+      'header-e',
+      'value',
+      'header-f',
+      'value'
+    ]
+    const arr1d = [
+      'header-c',
+      'value',
+      'header-a',
+      'value',
+      'header-b',
+      'value',
+      'header-5',
+      'value',
+      'header-e',
+      'value',
+      'header-d',
+      'value',
+      'header-f',
+      'value',
+      'header-1',
+      'value'
+    ]
+    const found = sort1d(arr1d)
+    const legacyFound = legacySort1d(arr1d)
+    t.strictDeepEqual(found, wanted)
+    t.strictDeepEqual(found, legacyFound)
   })
 })
